+++ src/utils/mod.rs
@@ -1,3 +1,4 @@
+pub mod debug;
 pub mod json;
 pub mod normalize;
 pub mod time;